remove common.js; restructure project; improve some code
