+++ utils/preview_gif.js
